Housekeeping, moving things to more sanely named packages.
