+++ sim-inventory/src/test/java/org/ostelco/simcards/EsimInventoryBatchImportIntegrationTest.kt
@@ -1,4 +1,4 @@
-package org.ostelco
+package org.ostelco.simcards
 
 import io.dropwizard.client.JerseyClientBuilder
 import io.dropwizard.jdbi.DBIFactory
